requirements.txt
streamlit>=1.32.0
streamlit-autorefresh>=1.0.1
requests>=2.31.0
py-clob-client>=0.14.0
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
httpx>=0.27.0
tenacity>=8.2.0
structlog>=24.1.0
cachetools>=5.3.0
rapidfuzz>=3.6.0
websockets>=12.0
pytest-asyncio>=0.23.0
